+++ src/Algorithms/ProximalGradientDescent.hpp
@@ -34,15 +34,15 @@ private:
     float tolerance;
     float learningRate;
     float learningRate2;
-    double prev_residue;
+    float prev_residue;
     long innerStep1;
     long innerStep2;
 
-    bool checkVectorConvergence(VectorXf, VectorXf, double);
+    bool checkVectorConvergence(VectorXf, VectorXf, float);
 
-    static constexpr double default_learning_rate = 0.001;
-    static constexpr double default_learning_rate2 = 0.001;
-    static constexpr double default_tolerance = 0.000001;
+    static constexpr float default_learning_rate = 0.001;
+    static constexpr float default_learning_rate2 = 0.001;
+    static constexpr float default_tolerance = 0.000001;
     static constexpr long default_inner_step1 = 10;
     static constexpr long default_inner_step2 = 10;
