+++ src/Models/LinearMixedModel.hpp
@@ -46,8 +46,8 @@ protected:
 
 //    MatrixXf beta; // d*1
     MatrixXf mau;  // Coeff matrix of similarity matrix.
-    double lambda_optimized; // Value at which log likelihood is max
-    double sigma;
+    float lambda_optimized; // Value at which log likelihood is max
+    float sigma;
     bool initFlag;
 
 public:
@@ -59,9 +59,9 @@ public:
     //Setters and Getters
     long get_num_samples();
     long get_X_features();
-    double get_lambda();
-    double getSigma();
-    void set_lambda(double);
+    float get_lambda();
+    float getSigma();
+    void set_lambda(float);
     void set_S(MatrixXf);
     void set_U(MatrixXf);
     void setXY(MatrixXf, MatrixXf);
@@ -70,14 +70,14 @@ public:
     MatrixXf getBeta();
     //Supporting functions
     // Final Objective of the LLM : Obtain beta matrix.
-    void calculate_beta(double);
-    void calculate_sigma(double);
-    double get_log_likelihood_value(double);
+    void calculate_beta(float);
+    void calculate_sigma(float);
+    float get_log_likelihood_value(float);
     void find_max_log_likelihood();
 
     void set_num_samples(int num_samples);
     // Search objective functions
-    double f(double);
+    float f(float);
     void assertReadyToRun();
     void init();
 };
